Migrate to maven-publishing Scala PR validation.
This commit only affects code paths where `$SCALA_REBUILD && $VALIDATOR`.
(Except for one slight tweak to preserve `MAVEN_ARGS` from outside.)

The new Scala PR validation publishes Scala to maven,
which is how downstream projects should consume it.
This makes reproducing builds much easier, as you don't need to
manually download any artifacts. If you have the jenkins artifactory in
your `~/.m2/settings.xml`, it's virtually transparent.

By relying on maven and a properties file that specifies the versions,
dependencies between the validation jobs are better encapsulated,
allowing easier evolution of the various integration builds.

The internal Maven repository is enabled for maven builds by
the new `pr-scala-integrate-ide`, which will pass `-P pr-scala` in `MAVEN_ARGS`.
(Which is why external `MAVEN_ARGS` are now preserved.)

The profile is configured on jenkins via `~/.m2/settings.xml`.

The URL to the repo is passed along to dbuild via the property `prRepoUrl`,
so that the `sbt-builds-for-ide` config can add it to the repositories used by dbuild.
The sbt launcher used for dbuild can't read environment vars
(or I don't know how to make it), so must pass it the Java way.

Eventually, the dbuild step should also leave the IDE validation script,
and be done as part of the main Scala PR validation, which integrates
all relevant modules anyway.